More accessible GS profile title and description.  Refs #9352.
Sorry, I should have been clearer about this in my review.  I think
we're trying to make sure that GS profile and descriptions are more
human accessible and require less technical/developer knowledge to
understand what they are.  Rather than bug Denys again, I've just
committed it myself, feel free to change it if I've gotten something
wrong.

Of course, this is largely irrelevant now that the profile is a
dependency of Plone's and has been hidden from the Plone UI as well
but my sense is that it's still best practice.

svn path=/plone.app.search/trunk/; revision=46246