Remove prevent default from change event handler
The change event handler on the advanced search to filter
types to search does have no effect. No other change handler
gets triggered.
But there is also a click handler.
On all Browsers except ie<=8, the click handler is called first.
On ie8, change events aren't bubbled. JQuery emulates the bubbling.
To do this, they created a click handler, that then simulates a
change handler.
The simulated change event triggers the change handler BEFORE
the click handler, the prevent default, prevents the
triggering of the click handler.
In the end, the type filter on checkboxes is never unchecked.
@pbauer