test_filter_empty: no batching in final results call.
We do no batching in the first few calls in this test
and it does not make sense (to me) to change that in the last one.
This test would sometimes pass and sometimes fail.
It could fail because somehow the batching returned my-page80 till my-page89
instead of the expected my-page0 till my-page9.
I do not know what causes that, but it seems irrelevant for this test.